# thesishub-application/api-docs/swagger.json
{
    "swagger":"2.0",
    "info":{
        "version":"1.1-SNAPSHOT",
        "title":"Thesis Hub Application"
    },
    "host":"localhost:8080",
    "basePath":"/",
    "schemes":[
        "http"
    ],
    "paths":{
...
        "/companies":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/Company"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            },
            "post":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/Company"
                        }
                    }
                ],
                "responses":{
                    "201":{
                        "description":"Created",
                        "headers":{
                            "Location":{
                                "type":"string"
                            }
                        },
                        "schema":{
                            "$ref":"#/definitions/Company"
                        }
                    },
                    "500":{
                        "description":"Internal Server Error",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/companies/{id}":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/Company"
                        }
                    }
                }
            },
            "put":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/Company"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/Company"
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            },
            "delete":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "204":{
                        "description":"No Content",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/companies/{id}/leaders":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/User"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/companies/{id}/plan":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/User"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/CompanyPlan"
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/companies/{id}/projects":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/Project"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/companies/{id}/topics":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/Topic"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
...
    },
    "definitions":{
        "Company":{
            "properties":{
                "city":{
                    "type":"string"
                },
                "country":{
                    "$ref":"#/definitions/Country"
                },
                "id":{
                    "type":"integer"
                },
                "logoUrl":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                },
                "plan":{
                    "$ref":"#/definitions/CompanyPlan"
                },
                "size":{
                    "type":"string",
                    "enum":[
                        "CORPORATE",
                        "MEDIUM",
                        "SMALL",
                        "STARTUP"
                    ]
                },
                "url":{
                    "type":"string"
                }
            }
        },
        "CompanyPlan":{
            "properties":{
                "description":{
                    "type":"string"
                },
                "maxTopics":{
                    "type":"integer"
                },
                "name":{
                    "type":"string"
                }
            }
        },
        "Country":{
            "properties":{
                "name":{
                    "type":"string"
                },
                "tag":{
                    "type":"string"
                }
            }
        },
        "Faculty":{
            "properties":{
                "id":{
                    "type":"integer"
                },
                "name":{
                    "type":"string"
                },
                "university":{
                    "$ref":"#/definitions/University"
                }
            }
        },
        "Project":{
            "properties":{
                "companies":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Company"
                    }
                },
                "creators":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/User"
                    }
                },
                "description":{
                    "type":"string"
                },
                "faculties":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Faculty"
                    }
                },
                "id":{
                    "type":"integer"
                },
                "name":{
                    "type":"string"
                },
                "topics":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Topic"
                    }
                }
            }
        },
...
        "Topic":{
            "properties":{
                "academicSupervisors":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/User"
                    }
                },
                "creator":{
                    "$ref":"#/definitions/User"
                },
                "degrees":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/TopicDegree"
                    }
                },
                "description":{
                    "type":"string"
                },
                "enabled":{
                    "type":"boolean"
                },
                "id":{
                    "type":"integer"
                },
                "secondaryDescription":{
                    "type":"string"
                },
                "secondaryTitle":{
                    "type":"string"
                },
                "shortAbstract":{
                    "type":"string"
                },
                "tags":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "title":{
                    "type":"string"
                }
            }
        },
...
        "TopicDegree":{
            "properties":{
                "description":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                }
            }
        },
        "University":{
            "properties":{
                "city":{
                    "type":"string"
                },
                "country":{
                    "$ref":"#/definitions/Country"
                },
                "id":{
                    "type":"integer"
                },
                "logoUrl":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                },
                "url":{
                    "type":"string"
                }
            }
        },
...
        "User":{
            "properties":{
                "company":{
                    "$ref":"#/definitions/Company"
                },
                "email":{
                    "type":"string"
                },
                "faculty":{
                    "$ref":"#/definitions/Faculty"
                },
                "id":{
                    "type":"integer"
                },
                "lastLogin":{
                    "type":"object"
                },
                "name":{
                    "type":"string"
                },
                "phone":{
                    "type":"string"
                },
                "roles":{
                    "type":"string",
                    "enum":[
                        "AC_SUPERVISOR",
                        "ADMIN",
                        "COMPANY_REP",
                        "PROJECT_LEADER",
                        "STUDENT",
                        "TECH_LEADER",
                        "UNIVERSITY_AMB"
                    ]
                },
                "tags":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "username":{
                    "type":"string"
                }
            }
        }
    }
}